More LOD and HUD stuff. Relflection is AAAAAAA
