+++ data/config.ini
@@ -112,5 +112,9 @@ DisplayRPM=1                # Display Vehicle RPM.
 AmberSignals=1              # 0 for original D2 Red, 1 for Amber orange turn signals
 FontChoice=1                # 0 for original, 1 for HD, 2 for Driver 1 HD 
 DigitChoice=1               # 0 for original, 1 for HD
-EnvLODDist=30000            # 0 ... 30k // Draw dist for LODs and TreeShadows. 
+
+#Extra DrawDistance (May Effect Performance) 30k is the edge of the 1800 drawDistance.
+EnvLODDist=30000            # 0 ... 30k // Draw dist for LODs. ReD2 default is 10,000 
 CarLODDIST=30000            # 0 ... 30k // Draw dist for Car LODs. 
+SpriteShadowDist= 30000     # 0 ... 30k // Draw Dist for Env Shadows. ReD2 default 14,000.
+GroundLODDist=12000         # 0 ... 12k // Road Tiles *Don't go over 12k please. Default ReD2 value is 7000
