+++ src_rebuild/Game/C/main.c
@@ -433,6 +433,7 @@ void LoadGameLevel(void)
 	SpoolLumpOffset = citylumps[GameLevel][CITYLUMP_SPOOL].x;
 
 	//Init_Reflection_Mapping();	// [A] I know that this is obsolete and used NOWHERE
+	//InitCarReflection();
 	InitDebrisNames();
 	InitShadow();
 	//InitTextureNames();			// [A] I know that this is obsolete and used NOWHERE
@@ -482,6 +483,7 @@ void State_GameInit(void* param)
 	InitEvents();
 	InitPadRecording();
 	InitSpeechQueue(&gSpeechQueue);
+	InitCarReflection();
 
 	if (NewLevel)
 	{
