+++ src_rebuild/Game/C/pres.c
@@ -19,40 +19,11 @@ extern int gShowMap;
 
 TextureID gHiresFontTexture = 0;
 TextureID gHiresDigitsTexture = 0;
-TextureID gCarReflectionTexture = 0;
 
 int gTextureChoice = 0; 
 int gDigitChoice = 0;
 
 stbtt_bakedchar gSTBCharData[224];	// ASCII 32..126 is 95 glyphs
-
-// Get Started
-void InitCarReflection()
-{
-	char namebuffer[64];
-	u_char* data;
-
-	// init Reflections Interactive
-	if (!gCarReflectionTexture)
-	{
-		int width, height, bpp;
-
-		
-		sprintf(namebuffer, "%s%s", gDataFolder, "GFX\\HQ\\REFLECTIONS.TGA");
-		
-
-		FS_FixPathSlashes(namebuffer);
-
-		if (LoadTGAImage(namebuffer, &data, width, height, bpp))
-		{
-			if (bpp == 32)
-			{
-				gCarReflectionTexture = GR_CreateRGBATexture(width, height, data);
-			}
-			free(data);
-		}
-	}
-}
 
 void InitHiresFonts()
 {
@@ -126,27 +97,6 @@ void InitHiresFonts()
 			free(data);
 		}
 	}
-}
-
-// [A] attempt to restore D1 reflections on car.
-// Not working yet.
-void SetCarReflection(int enabled)
-{
-
-	if (gCarReflectionTexture == 0)
-	{
-		return;
-	}
-
-	DR_PSYX_TEX* tex = (DR_PSYX_TEX*)current->primptr;
-	if (enabled)
-		SetPsyXTexture(tex, gCarReflectionTexture, 255, 255);
-	else
-		SetPsyXTexture(tex, 0, 0, 0);
-
-	addPrim(current->ot, tex);
-	current->primptr += sizeof(DR_PSYX_TEX);
-
 }
 
 void SetHiresFontTexture(int enabled)
